use fen + move as card id
The anki card id must be an uniq value.
We previously use the fen of the board of this card question as id.
This doesn't work because for a specific board, many opponent moves can
be possible. For example, in the french defense, regarding the lichess
opening, after "1.e4 e6", we could have "2.d4", "2.f3" or "3.c3".

To avoid these duplicate ids we compute the card id with the md5sum of
the fen and opponent san move